article_details

Retrieves data about a single paper and all of its authors

article_details(ID)

Arguments

ID
Rxivist paper ID.

Value

Returns a list with details about the specified article.

Examples

article_details(72514)

#> $id
#> [1] "72514"
#> 
#> $doi
#> [1] "10.1101/2020.01.30.927871"
#> 
#> $first_posted
#> [1] "2020-01-31"
#> 
#> $biorxiv_url
#> [1] "https://www.biorxiv.org/content/10.1101/2020.01.30.927871v2"
#> 
#> $url
#> [1] "https://api.rxivist.org/v1/papers/72514"
#> 
#> $title
#> [1] "Uncanny similarity of unique inserts in the 2019-nCoV spike protein to HIV-1 gp120 and Gag"
#> 
#> $category
#> [1] "evolutionary-biology"
#> 
#> $abstract
#> [1] "This paper has been withdrawn by its authors. They intend to revise it in response to comments received from the research community on their technical approach and their interpretation of the results. If you have any questions, please contact the corresponding author."
#> 
#> $authors
#>       id                     name
#> 1 580441         Prashant Pradhan
#> 2 580442    Ashutosh Kumar Pandey
#> 3 580443          Akhilesh Mishra
#> 4 580444              Parul Gupta
#> 5 580445   Praveen Kumar Tripathi
#> 6 582554 Manoj Balakrishnan Menon
#> 7 295517              James Gomes
#> 8 580447     Perumal Vivekanandan
#> 9 580448          Bishwajit Kundu
#>                                                            institution orcid
#> 1                                 Indian Institute of Technology Delhi    NA
#> 2                                 Indian Institute of Technology Delhi    NA
#> 3                            Indian Institute of Technology, New Delhi    NA
#> 4                                 Indian Institute of Technology Delhi    NA
#> 5                                 Indian Institute of Technology Delhi    NA
#> 6 Kusuma School of biological sciences, Indian institute of technology    NA
#> 7                                 Indian Institute of Technology Delhi    NA
#> 8                                 Indian Institute of Technology Delhi    NA
#> 9                                 Indian Institute of Technology Delhi    NA
#> 
#> $ranks
#> $ranks$alltime
#> $ranks$alltime$downloads
#> [1] 962296
#> 
#> $ranks$alltime$rank
#> [1] 1
#> 
#> $ranks$alltime$out_of
#> [1] 94912
#> 
#> $ranks$alltime$tie
#> [1] FALSE
#> 
#> 
#> $ranks$ytd
#> $ranks$ytd$downloads
#> [1] 962296
#> 
#> $ranks$ytd$rank
#> [1] 1
#> 
#> $ranks$ytd$out_of
#> [1] 94912
#> 
#> $ranks$ytd$tie
#> [1] FALSE
#> 
#> 
#> $ranks$lastmonth
#> $ranks$lastmonth$downloads
#> [1] 14987
#> 
#> $ranks$lastmonth$rank
#> [1] 3
#> 
#> $ranks$lastmonth$out_of
#> [1] 94912
#> 
#> $ranks$lastmonth$tie
#> [1] FALSE
#> 
#> 
#> $ranks$category
#> $ranks$category$downloads
#> [1] 962296
#> 
#> $ranks$category$rank
#> [1] 1
#> 
#> $ranks$category$out_of
#> [1] 5736
#> 
#> $ranks$category$tie
#> [1] FALSE
#> 
#> 
#> 
#> $publication
#> named list()
#>