author

Provides information about the specified author.

author(ID)

Arguments

ID
Rxivist author ID.

Value

Returns a data frame with information about the specified author.

Examples

author(206102)

#> $id
#> [1] 206102
#> 
#> $name
#> [1] "Koen Van Waerebeek"
#> 
#> $institution
#> [1] "Peruvian Centre for Cetacean Research"
#> 
#> $orcid
#> NULL
#> 
#> $emails
#> [1] "corewam@yandex.com"
#> 
#> $articles
#>      id            doi                                      biorxiv_url
#> 1 17776 10.1101/042614 https://www.biorxiv.org/content/10.1101/042614v1
#> 2 17779 10.1101/035337 https://www.biorxiv.org/content/10.1101/035337v1
#> 3 17735 10.1101/185033 https://www.biorxiv.org/content/10.1101/185033v1
#> 4 38716 10.1101/101915 https://www.biorxiv.org/content/10.1101/101915v1
#>                                       url
#> 1 https://api.rxivist.org/v1/papers/17776
#> 2 https://api.rxivist.org/v1/papers/17779
#> 3 https://api.rxivist.org/v1/papers/17735
#> 4 https://api.rxivist.org/v1/papers/38716
#>                                                                                                                                                            title
#> 1        Omura's whale (Balaenoptera omurai) stranding on Qeshm Island, Iran: further evidence for a wide (sub)tropical distribution, including the Persian Gulf
#> 2            New records of Atlantic humpback dolphin in Guinea, Nigeria, Cameroon and Togo underscore fisheries pressure and generalised marine bushmeat demand
#> 3                                                                                                                            Humpback whales of the Persian Gulf
#> 4 Sex differences in the epidemiology of tattoo skin disease in captive common bottlenose dolphins (Tursiops truncatus): are males more vulnerable than females?
#>       category ranks.alltime.downloads ranks.alltime.rank ranks.alltime.out_of
#> 1      zoology                    1537               6036                94912
#> 2      zoology                    1112              10346                94912
#> 3      zoology                     723              20243                94912
#> 4 epidemiology                     473              35738                94912
#>   ranks.alltime.tie ranks.ytd.downloads ranks.ytd.rank ranks.ytd.out_of
#> 1             FALSE                  66          71252            94912
#> 2             FALSE                  96          51254            94912
#> 3             FALSE                 169          26799            94912
#> 4             FALSE                  79          61687            94912
#>   ranks.ytd.tie ranks.lastmonth.downloads ranks.lastmonth.rank
#> 1         FALSE                        11                60307
#> 2         FALSE                        16                45317
#> 3         FALSE                        39                18480
#> 4         FALSE                        14                51085
#>   ranks.lastmonth.out_of ranks.lastmonth.tie ranks.category.downloads
#> 1                  94912               FALSE                     1537
#> 2                  94912               FALSE                     1112
#> 3                  94912               FALSE                      723
#> 4                  94912               FALSE                      473
#>   ranks.category.rank ranks.category.out_of ranks.category.tie
#> 1                  13                   494              FALSE
#> 2                  29                   494              FALSE
#> 3                  65                   494              FALSE
#> 4                 506                  1556              FALSE
#> 
#> $ranks
#>   downloads  rank out_of   tie     category
#> 1      3845 27808 404161  TRUE      alltime
#> 2       473  4078   9545  TRUE epidemiology
#> 3      3372     1   2420 FALSE      zoology
#>