Authors ranks

Returns top 200 authors in specified category.

authors_rank(category = "")

Arguments

category
Filters out results not related to the specified category. Category list can be viewed with category_list function. If filtering for a single category, this category can be specified as a string. If filtering for multiple categories, the argument must be specified as a vector, e.g. category = c("zoology","biophysics"). If it is left as default, the function will return top 200 authors overall.

Value

Returns a data frame with the top 200 authors.

Examples

authors_rank("plant-biology")

#>         id                               name rank downloads   tie
#> 1   202924                     Sophien Kamoun    1     30805 FALSE
#> 2   205706                 Daniel H. Chitwood    2     24729 FALSE
#> 3   214840                     Cristobal Uauy    3     22683 FALSE
#> 4   496223                           Y. Yovel    4     21453  TRUE
#> 5   496222                         A. Boonman    4     21453  TRUE
#> 6   496219                           I. Khait    4     21453  TRUE
#> 7   543993                           K. Saban    4     21453  TRUE
#> 8   224115                 Ohad Lewin-Epstein    4     21453  TRUE
#> 9   496221                        R. Perelman    4     21453  TRUE
#> 10  496224                          L. Hadany    4     21453  TRUE
#> 11  496220                          R. Sharon    4     21453  TRUE
#> 12  216391                     James Simmonds   12     20174 FALSE
#> 13  210387                      Wendy Harwood   13     18785 FALSE
#> 14  205643                            Ji Zhou   14     17758 FALSE
#> 15  597982                     Paul Nicholson   15     17639  TRUE
#> 16  226921                      Lee T. Hickey   15     17639  TRUE
#> 17  226899                        Sreya Ghosh   15     17639  TRUE
#> 18  226900                         Amy Watson   15     17639  TRUE
#> 19  550038                 Brande B. H. Wulff   15     17639  TRUE
#> 20  226904                      Tracey Rayner   15     17639  TRUE
#> 21  226909                       Andrew Steed   15     17639  TRUE
#> 22  226918                     Claire Domoney   15     17639  TRUE
#> 23  205644                    Daniel Reynolds   23     16563 FALSE
#> 24  214654                   Brian Staskawicz   24     13688 FALSE
#> 25  340486                     Andrey Korolev   25     13682 FALSE
#> 26  402199                   Siobhan M. Brady   26     13446 FALSE
#> 27  220659                       Graham Moore   27     13089  TRUE
#> 28  220658                  María-Dolores Rey   27     13089  TRUE
#> 29  210371                M. Asyraf Md. Hatta   29     13013 FALSE
#> 30  340492                        Raman Harsh   30     12979 FALSE
#> 31  606296                      Robert F Park   31     12523 FALSE
#> 32  340493                   Christian Rogers   32     12431 FALSE
#> 33  278886                      David Edwards   33     12425  TRUE
#> 34  340491                  Jacqueline Batley   33     12425  TRUE
#> 35  340496                     Scott A. Boden   35     12267 FALSE
#> 36  340489                     William Martin   36     12172  TRUE
#> 37  340487                     Laura E. Dixon   36     12172  TRUE
#> 38  340490                       Merrill Ryan   36     12172  TRUE
#> 39  340485                    Andy Breakspear   36     12172  TRUE
#> 40  340494                    Mark J. Dieters   36     12172  TRUE
#> 41  340483                  Alison Hinchliffe   36     12172  TRUE
#> 42  340488                         Adnan Riaz   36     12172  TRUE
#> 43  340495                      Ian H. DeLacy   36     12172  TRUE
#> 44  340484                    Nikolai Adamski   36     12172  TRUE
#> 45  428381                 Matthew J Williams   36     12172  TRUE
#> 46  292505                   William S. Cuddy   36     12172  TRUE
#> 47  214665                   Tolga O. Bozkurt   47     11439 FALSE
#> 48  200333                        Ivan Baxter   48     11262 FALSE
#> 49  218625                      Detlef Weigel   49     10432 FALSE
#> 50  219268                      Alexis Maizel   50      9990 FALSE
#> 51  213403                     Jan U. Lohmann   51      9900 FALSE
#> 52  214655                  Markita P. Landry   52      9390  TRUE
#> 53  214644                   Gozde S. Demirer   52      9390  TRUE
#> 54  214645                         Huan Zhang   52      9390  TRUE
#> 55  213310                  Jonathan DG Jones   55      9291 FALSE
#> 56  205705                             Mao Li   56      9279 FALSE
#> 57  205633                     Khaoula Belhaj   57      8892 FALSE
#> 58  216501                       Dolf Weijers   58      8817 FALSE
#> 59  205628                       Chih-Hang Wu   59      8787 FALSE
#> 60  214653                      Myeong-Je Cho   60      8624 FALSE
#> 61  214650                        Roger Chang   61      8550 FALSE
#> 62  223296                        Mark Tester   62      8497 FALSE
#> 63  210471                      Roger B. Deal   63      8476 FALSE
#> 64  218055                    Juliana L Matos   64      8415 FALSE
#> 65  214651                 Abhishek J Aditham   65      8379 FALSE
#> 66  210533                  Anton Sharakshane   66      8134 FALSE
#> 67  430440                       Shou-Ling Xu   67      8108 FALSE
#> 68  210358                    Daniela Vergara   68      8076 FALSE
#> 69  210502                   Margaret H Frank   69      8034 FALSE
#> 70  205702                       Cyril Zipfel   70      7888 FALSE
#> 71  332664                  Matthew J. Moscou   71      7843 FALSE
#> 72  200908                     Jeremy Schmutz   72      7827 FALSE
#> 73  216412                       Blake Meyers   73      7807 FALSE
#> 74  212093                  Julian M. Hibberd   74      7798 FALSE
#> 75  205637                   Mark J. Banfield   75      7792 FALSE
#> 76  209475                Paul Schulze-Lefert   76      7752 FALSE
#> 77  210365                      Donald P Land   77      7674  TRUE
#> 78  210363                Travis C Ruthenburg   77      7674  TRUE
#> 79  210366                       Nolan C Kane   77      7674  TRUE
#> 80  210364                    Kymron deCesare   77      7674  TRUE
#> 81  213397                    Christian Wenzl   81      7566 FALSE
#> 82  214649                      Younghun Sung   82      7539  TRUE
#> 83  214652                         Linda Chio   82      7539  TRUE
#> 84  214647                        Natalie Goh   82      7539  TRUE
#> 85  499167                 Francis Cunningham   82      7539  TRUE
#> 86  210418                    Guillaume Lobet   86      7515 FALSE
#> 87  221539                  Michael J. Axtell   87      7425 FALSE
#> 88  247889                   Benjamin J. Cole   88      7334 FALSE
#> 89  249436                       Alice Y Ting   89      7295 FALSE
#> 90  205636                    Ryohei Terauchi   90      7236 FALSE
#> 91  214630                       Rachel Wells   91      7222 FALSE
#> 92  217979                 Alisdair R. Fernie   92      7120 FALSE
#> 93  222534                   Viktoriya Coneva   93      7032 FALSE
#> 94  221505                    Nathan Springer   94      7005 FALSE
#> 95  211951                Diane G.O. Saunders   95      6862 FALSE
#> 96  221562                         Jiří Friml   96      6763 FALSE
#> 97  210615                Sebastian Schornack   97      6743 FALSE
#> 98  458939           Ricardo Ramírez-González   98      6736 FALSE
#> 99  211952                            Joe Win   99      6659 FALSE
#> 100 529351                    Thomas A. Swift  100      6605  TRUE
#> 101 529356                    M. Carmen Galan  100      6605  TRUE
#> 102 529357                 Heather M. Whitney  100      6605  TRUE
#> 103 529354              David Benito-Alifonso  100      6605  TRUE
#> 104 211990                Julia Bailey-Serres  104      6586 FALSE
#> 105 210565                        Thomas Greb  105      6570 FALSE
#> 106 222511                     Wolfgang Busch  106      6568 FALSE
#> 107 522746                    R Keith Slotkin  107      6482 FALSE
#> 108 202817                  Robert J. Schmitz  108      6459 FALSE
#> 109 210359                       Silas Tittes  109      6412  TRUE
#> 110 210357                       Ryan C Lynch  109      6412  TRUE
#> 111 210362                    Matthew J Gibbs  109      6412  TRUE
#> 112 210361                      C.J. Schwartz  109      6412  TRUE
#> 113 210360                      Kristin White  109      6412  TRUE
#> 114 220651                    Michael Hothorn  114      6309 FALSE
#> 115 219341                       Jing-Ke Weng  115      6231 FALSE
#> 116 529349                         Cara Doyle  116      6227  TRUE
#> 117 529352                      Mark Winfield  116      6227  TRUE
#> 118 529350                 Katie Higginbottom  116      6227  TRUE
#> 119 529355                     Taryn Fletcher  116      6227  TRUE
#> 120 323966                      Keith Edwards  116      6227  TRUE
#> 121 529353                 Christopher Bellas  116      6227  TRUE
#> 122 210501                 Jürgen Kleine-Vehn  122      6126 FALSE
#> 123 226905                         Phon Green  123      6101 FALSE
#> 124 202441                    Joseph R. Ecker  124      5986 FALSE
#> 125 203519                 Christine Queitsch  125      5907 FALSE
#> 126 205704                     Zoë Migicovsky  126      5862 FALSE
#> 127 526852              Aleksandra S Tsarkova  127      5779  TRUE
#> 128 526856                    Pavel V Shalaev  127      5779  TRUE
#> 129 252379                  Ilia V. Yampolsky  127      5779  TRUE
#> 130 526859                     Sergey M Deyev  127      5779  TRUE
#> 131 526860                   Andrey I Bubyrev  127      5779  TRUE
#> 132 526849                  Kseniia A Palkina  127      5779  TRUE
#> 133 526846         Louisa Gonzalez Somermeyer  127      5779  TRUE
#> 134 526857                    Olesya A Melnik  127      5779  TRUE
#> 135 526845               Tatiana Mitiouchkina  127      5779  TRUE
#> 136 499346                 Karen S. Sarkisyan  127      5779  TRUE
#> 137 252376               Ekaterina S Shakhova  127      5779  TRUE
#> 138 526858               Victoria O Shipunova  127      5779  TRUE
#> 139 526861                 Alexander S Pushin  127      5779  TRUE
#> 140 526862                   Vladimir V Choob  127      5779  TRUE
#> 141 252375                Tatiana V Chepurnyh  127      5779  TRUE
#> 142 526851                    Sofia V Chekova  127      5779  TRUE
#> 143 526848                Tatiana A Karataeva  127      5779  TRUE
#> 144 526853                 Yaroslav V Golubev  127      5779  TRUE
#> 145 332551               Fyodor A. Kondrashov  127      5779  TRUE
#> 146 526847                     Elena B Guglya  127      5779  TRUE
#> 147 526863                    Sergey V Dolgov  127      5779  TRUE
#> 148 526850              Liliia I Fakhranurova  127      5779  TRUE
#> 149 500918                Alexander S. Mishin  127      5779  TRUE
#> 150 526855                 Sergey A Dolgushin  127      5779  TRUE
#> 151 252374                 Nadezhda M Markina  127      5779  TRUE
#> 152 526854                Vadim V Negrebetsky  127      5779  TRUE
#> 153 210556                        Marko Bajic  153      5748 FALSE
#> 154 213302                        Kamil Witek  154      5715 FALSE
#> 155 214613                     Benjamin Petre  155      5699 FALSE
#> 156 205668                  James C. Schnable  156      5695 FALSE
#> 157 218056              Dominique C. Bergmann  157      5677 FALSE
#> 158 216404                    José R. Dinneny  158      5649 FALSE
#> 159 222601                    Todd C. Mockler  159      5635 FALSE
#> 160 214663                        Cian Duggan  160      5596 FALSE
#> 161 205625                    Kentaro Yoshida  161      5590 FALSE
#> 162 214719                  Dmitri A. Nusinow  162      5561 FALSE
#> 163 216553                       Marcel Quint  163      5554 FALSE
#> 164 211908                   Neelima R. Sinha  164      5545 FALSE
#> 165 214685                         M. Estelle  165      5477 FALSE
#> 166 221508                   Philip N. Benfey  166      5468 FALSE
#> 167 226919                    Brittany Hazard  167      5467  TRUE
#> 168 695625                     Cristobal Uauy  167      5467  TRUE
#> 169 226910                   Abhimanyu Sarkar  167      5467  TRUE
#> 170 226903             Marcela Mendoza-Suárez  167      5467  TRUE
#> 171 226913                          John Lord  167      5467  TRUE
#> 172 226912                     Lionel Perkins  167      5467  TRUE
#> 173 226907                       Sadiye Hayta  167      5467  TRUE
#> 174 226902                         Luis Yanes  167      5467  TRUE
#> 175 226911                      Jeremy Carter  167      5467  TRUE
#> 176 226901          Oscar E. Gonzalez-Navarro  167      5467  TRUE
#> 177 226906                       Amber Hafeez  167      5467  TRUE
#> 178 226908                   Rachel E. Melton  167      5467  TRUE
#> 179 564113                      Cathie Martin  167      5467  TRUE
#> 180 226914                       Anne Osbourn  167      5467  TRUE
#> 181 350973 Daniela Paula de Toledo Thomazella  181      5354  TRUE
#> 182 350974                      Quinton Brail  181      5354  TRUE
#> 183 350975                   Douglas Dahlbeck  181      5354  TRUE
#> 184 213386                     Malia A. Gehan  184      5334 FALSE
#> 185 211982                       Kaisa Kajala  185      5264 FALSE
#> 186 211947                  Bruce A. McDonald  186      5261 FALSE
#> 187 268031                         I Feussner  187      5234 FALSE
#> 188 339973                       Teva Vernoux  188      5147 FALSE
#> 189 213399                Amaya Vilches Barro  189      5138 FALSE
#> 190 200910                      Jane Grimwood  190      5133 FALSE
#> 191 219288               Rubén Rellán-Álvarez  191      5107 FALSE
#> 192 480860                    Cheng-Bin Xiang  192      5089 FALSE
#> 193 375547              Krzysztof Zienkiewicz  193      5052 FALSE
#> 194 226670                     Alain Goossens  194      5018 FALSE
#> 195 206412                    Bart PHJ Thomma  195      4988 FALSE
#> 196 211912                         Remco Stam  196      4964 FALSE
#> 197 223549                        Jan Sklenar  197      4948 FALSE
#> 198 226643                        Tom Bennett  198      4939 FALSE
#> 199 430285                        Tae-Wuk Kim  199      4926  TRUE
#> 200 430286                       Chan Ho Park  199      4926  TRUE