Search Rxivist

Retrieves a articles with matching description

rxivist_search(
  search_phrase = "",
  from = "alltime",
  category = "",
  sortby = "downloads",
  limit = 20
)

Arguments

search_phrase
Search phrase used for finding articles. Can be left empty.

from
Defines what timeframe to search. Can be alltime, ytd, lastmonth, day, week, month, year. Lastmonth and ytd can only be used with sortby set as downloads

category
Filters out results not related to the specified category. Category list can be viewed with category_list function. If filtering for a single category, this category can be specified as a string. If filtering for multiple categories, argument must be specified as a vector, e.g. category = c("zoology","biophysics").

sortby
Sorts the results based on the number of downloads or twitter statistics.

limit
Number of results to retrieve.

Value

Returns a data frame with articles matching specified search criteria.

Examples

rxivist_search(search_phrase = "T-cells", from = "alltime", sortby = "twitter", limit = 5)

#>      id metric
#> 1 72514   9806
#> 2 71702   3947
#> 3 52089   3175
#> 4 72040   3158
#> 5 55953   2374
#>                                                                                                                   title
#> 1                            Uncanny similarity of unique inserts in the 2019-nCoV spike protein to HIV-1 gp120 and Gag
#> 2 Discovery of a novel coronavirus associated with the recent pneumonia outbreak in humans and its potential bat origin
#> 3                                                       7 Tesla MRI of the ex vivo human brain at 100 micron resolution
#> 4                                            Time-varying transmission dynamics of Novel Coronavirus Pneumonia in China
#> 5                                             An integrated brain-machine interface platform with thousands of channels
#>                                       url
#> 1 https://api.rxivist.org/v1/papers/72514
#> 2 https://api.rxivist.org/v1/papers/71702
#> 3 https://api.rxivist.org/v1/papers/52089
#> 4 https://api.rxivist.org/v1/papers/72040
#> 5 https://api.rxivist.org/v1/papers/55953
#>                                                   biorxiv_url
#> 1 https://www.biorxiv.org/content/10.1101/2020.01.30.927871v2
#> 2 https://www.biorxiv.org/content/10.1101/2020.01.22.914952v2
#> 3            https://www.biorxiv.org/content/10.1101/649822v2
#> 4 https://www.biorxiv.org/content/10.1101/2020.01.25.919787v2
#> 5            https://www.biorxiv.org/content/10.1101/703801v4
#>                         doi             category first_posted
#> 1 10.1101/2020.01.30.927871 evolutionary-biology   2020-01-31
#> 2 10.1101/2020.01.22.914952         microbiology   2020-01-23
#> 3            10.1101/649822         neuroscience   2019-05-31
#> 4 10.1101/2020.01.25.919787      systems-biology   2020-01-26
#> 5            10.1101/703801         neuroscience   2019-07-17
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    abstract
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              This paper has been withdrawn by its authors. They intend to revise it in response to comments received from the research community on their technical approach and their interpretation of the results. If you have any questions, please contact the corresponding author.
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              Since the SARS outbreak 18 years ago, a large number of severe acute respiratory syndrome related coronaviruses (SARSr-CoV) have been discovered in their natural reservoir host, bats. Previous studies indicated that some of those bat SARSr-CoVs have the potential to infect humans. Here we report the identification and characterization of a novel coronavirus (nCoV-2019) which caused an epidemic of acute respiratory syndrome in humans, in Wuhan, China. The epidemic, started from December 12th, 2019, has caused 198 laboratory confirmed infections with three fatal cases by January 20th, 2020. Full-length genome sequences were obtained from five patients at the early stage of the outbreak. They are almost identical to each other and share 79.5% sequence identify to SARS-CoV. Furthermore, it was found that nCoV-2019 is 96% identical at the whole genome level to a bat coronavirus. The pairwise protein sequence analysis of seven conserved non-structural proteins show that this virus belongs to the species of SARSr-CoV. The nCoV-2019 virus was then isolated from the bronchoalveolar lavage fluid of a critically ill patient, which can be neutralized by sera from several patients. Importantly, we have confirmed that this novel CoV uses the same cell entry receptor, ACE2, as SARS-CoV.
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                We present an ultra-high resolution MRI dataset of an ex vivo human brain specimen. The brain specimen was donated by a 58-year-old woman who had no history of neurological disease and died of non-neurological causes. After fixation in 10% formalin, the specimen was imaged on a 7 Tesla MRI scanner at 100 μm isotropic resolution using a custom-built 31-channel receive array coil. Single-echo multi-flip Fast Low-Angle SHot (FLASH) data were acquired over 100 hours of scan time (25 hours per flip angle), allowing derivation of a T1 parameter map and synthesized FLASH volumes. This dataset provides an unprecedented view of the three-dimensional neuroanatomy of the human brain. To optimize the utility of this resource, we warped the dataset into standard stereotactic space. We now distribute the dataset in both native space and stereotactic space to the academic community via multiple platforms. We envision that this dataset will have a broad range of investigational, educational, and clinical applications that will advance understanding of human brain anatomy in health and disease.\n\nView this table:
#> 4 Rationale Several studies have estimated basic production number of novel coronavirus pneumonia (NCP). However, the time-varying transmission dynamics of NCP during the outbreak remain unclear.\n\nObjectives We aimed to estimate the basic and time-varying transmission dynamics of NCP across China, and compared them with SARS.\n\nMethods Data on NCP cases by February 7, 2020 were collected from epidemiological investigations or official websites. Data on severe acute respiratory syndrome (SARS) cases in Guangdong Province, Beijing and Hong Kong during 2002-2003 were also obtained. We estimated the doubling time, basic reproduction number ( R ) and time-varying reproduction number ( Rt ) of NCP and SARS.\n\nMeasurements and main results As of February 7, 2020, 34,598 NCP cases were identified in China, and daily confirmed cases decreased after February 4. The doubling time of NCP nationwide was 2.4 days which was shorter than that of SARS in Guangdong (14.3 days), Hong Kong (5.7 days) and Beijing (12.4 days). The R of NCP cases nationwide and in Wuhan were 4.5 and 4.4 respectively, which were higher than R of SARS in Guangdong ( R =2.3), Hongkong ( R =2.3), and Beijing ( R =2.6). The Rt for NCP continuously decreased especially after January 16 nationwide and in Wuhan. The R for secondary NCP cases in Guangdong was 0.6, and the Rt values were less than 1 during the epidemic.\n\nConclusions NCP may have a higher transmissibility than SARS, and the efforts of containing the outbreak are effective. However, the efforts are needed to persist in for reducing time-varying reproduction number below one.\n\nScientific Knowledge on the Subject Since December 29, 2019, pneumonia infection with 2019-nCoV, now named as Novel Coronavirus Pneumonia (NCP), occurred in Wuhan, Hubei Province, China. The disease has rapidly spread from Wuhan to other areas. As a novel virus, the time-varying transmission dynamics of NCP remain unclear, and it is also important to compare it with SARS.\n\nWhat This Study Adds to the Field We compared the transmission dynamics of NCP with SARS, and found that NCP has a higher transmissibility than SARS. Time-varying production number indicates that rigorous control measures taken by governments are effective across China, and persistent efforts are needed to be taken for reducing instantaneous reproduction number below one.
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        Brain-machine interfaces (BMIs) hold promise for the restoration of sensory and motor function and the treatment of neurological disorders, but clinical BMIs have not yet been widely adopted, in part because modest channel counts have limited their potential. In this white paper, we describe Neuralink’s first steps toward a scalable high-bandwidth BMI system. We have built arrays of small and flexible electrode “threads”, with as many as 3,072 electrodes per array distributed across 96 threads. We have also built a neurosurgical robot capable of inserting six threads (192 electrodes) per minute. Each thread can be individually inserted into the brain with micron precision for avoidance of surface vasculature and targeting specific brain regions. The electrode array is packaged into a small implantable device that contains custom chips for low-power on-board amplification and digitization: the package for 3,072 channels occupies less than (23 × 18.5 × 2) mm3. A single USB-C cable provides full-bandwidth data streaming from the device, recording from all channels simultaneously. This system has achieved a spiking yield of up to 70% in chronically implanted electrodes. Neuralink’s approach to BMI has unprecedented packaging density and scalability in a clinically relevant package.
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    authors
#> 1                                                                                                                                                                                                                                                                                                                                                      580441, 580442, 580443, 580444, 580445, 582554, 295517, 580447, 580448, Prashant Pradhan, Ashutosh Kumar Pandey, Akhilesh Mishra, Parul Gupta, Praveen Kumar Tripathi, Manoj Balakrishnan Menon, James Gomes, Perumal Vivekanandan, Bishwajit Kundu
#> 2 331721, 576248, 576249, 493722, 554523, 375415, 576250, 338835, 235857, 576251, 576252, 207306, 698952, 576253, 576254, 576255, 231411, 576256, 203382, 576257, 273173, 576258, 466252, 576259, 576260, 576261, 576262, 294673, 576247, Peng Zhou, Xing-Lou Yang, Xian-Guang Wang, Ben Hu, Lei Zhang, Wei Zhang, Hao-Rui Si, Yan Zhu, Bei Li, Chao-Lin Huang, Hui-Dong Chen, Jing Chen, Yun Luo, Hua Guo, Ren-Di Jiang, Mei-Qin Liu, Ying Chen, Xu-Rui Shen, Xi Wang, Xiao-Shuang Zheng, Kai Zhao, Quan-Jiao Chen, Fei Deng, Lin-Lin Liu, Bing Yan, Fa-Xian Zhan, Yan-Yi Wang, Gengfu Xiao, Zheng-Li Shi
#> 3                                                                                                                                                                                                                      445770, 445771, 229785, 367435, 462235, 279299, 445773, 477846, 445775, 308363, 229297, 445776, 365812, 217897, 367436, Brian L. Edlow, Azma Mareyam, A Horn, Jonathan R. Polimeni, Thomas Witzel, M. Dylan Tisdall, Jean Augustinack, Jason P. Stockmann, Bram R. Diamond, Allison Stevens, Lee S. Tirrell, Rebecca D Folkerth, Lawrence L Wald, Bruce Fischl, Andre van der Kouwe
#> 4                                                                               207996, 577925, 577927, 577928, 200441, 577931, 309640, 577926, 577930, 577932, 577933, 577934, 577935, 577936, 577937, 586904, 577938, 586905, 586906, 211265, 586907, 586908, 577929, 577940, 524666, Tao Liu, Jianxiong Hu, Jianpeng Xiao, Guanhao He, Min Kang, Zuhua Rong, Lifeng Lin, Haojie Zhong, Qiong Huang, Aiping Deng, Weilin Zeng, Xiaohua Tan, Siqing Zeng, Zhihua Zhu, Jiansen Li, Dexin Gong, Donghua Wan, Shaowei Chen, Lingchuan Guo, Yan Li, Limei Sun, Wenjia Liang, Tie Song, Jianfeng He, Wenjun Ma
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     461787, 461788, Elon Musk, Neuralink